==> analyses/tp53_nf1_score/05-tp53-altered-annotation.Rmd <==
---
title: "Tp53 SNV hotspots"
author: "K S Gaonkar, Jo Lynne Rokita"
output: html_notebook
params:
  base_run:
    label: "1/0 to run with base histology"
    value: 0
    input: integer
---

In this notebook we will add TP53 alteration status as discussed in [#837](https://github.com/AlexsLemonade/OpenPBTA-analysis/issues/837):

**TP53 altered - loss**, if :

 - A sample contains a TP53 hotspot SNV mutation. (Cancer hotspot database and downloadable file available). Please also crosscheck that all mutations from this table are included. Source: https://cavatica.sbgenomics.com/u/kfdrc-harmonization/kf-reference-pipeline/files/#q?search=hot
 - A sample contains two TP53 alterations, suggesting (but not confirming) that both alleles are affected (SNV+SNV, (CNV or SV) + SNV).
 - A sample contains one alteration (SNV or (CNV or SV)) + has cancer_predispositions == "Li-Fraumeni syndrome", suggesting there is a germline variant in addition to the somatic variant we observe.
 - A sample does not have a TP53 alterations, but has cancer_predispositions == "Li-Fraumeni syndrome" and TP53 classifier score for matched RNA-Seq > 0.5 (or higher cutoff we decide upon later).
 
 Note: CNV and SV will be considered as the same event, but we will be adding SV_counts and SV.type to the altered status output file
 
**TP53 altered - activated**, if:

- A sample contains one of the two TP53 activating mutations p.R273C and p.R248W. [@doi:10.1038/ng0593-42]

## Setup
```{r read in histology, make match id}
library("ggpubr")
library("ggthemes")
library("tidyverse")
library("ggfortify")
library("broom")

# rootdir
root_dir <- rprojroot::find_root(rprojroot::has_dir(".git"))
data_dir <- file.path(root_dir, "data")

input_dir <- file.path(root_dir,
                         "analyses",
                         "tp53_nf1_score",
                         "input")
# histology
if ( params$base_run ==0 ){
  clinical<-read.delim(file.path(data_dir,"histologies.tsv"), stringsAsFactors = FALSE)
} else{
  clinical<-read.delim(file.path(data_dir,"histologies-base.tsv"), stringsAsFactors = FALSE)  
}

histology <- clinical %>%
  # remove TCGA, GTEX from altered file since we don't have DNA
  filter(!is.na(pathology_diagnosis),
         !cohort %in% c("TCGA", "GTEx")) %>%
  dplyr::select(Kids_First_Biospecimen_ID, match_id, cancer_predispositions,
  experimental_strategy) %>%
    unique()

# Cancer database
hotspot_database_2017_tp53_snv <- read_tsv(file.path(input_dir,"protein_snv_cancer_hotspots_v2.ENS105_liftover.tsv")) %>%
  filter(Hugo_Symbol == "TP53")
hotspot_database_2017_tp53_indel <- read_tsv(file.path(input_dir,"protein_indel_cancer_hotspots_v2.ENS105_liftover.tsv")) %>%
  filter(Hugo_Symbol == "TP53")

# p53, p63 and p73 functional sites
functional_sites_tp53 <- read_tsv(file.path(input_dir,"hotspot_Chen_2006.tsv")) 

results_dir <- file.path(root_dir,
                         "analyses",
                         "tp53_nf1_score",
                         "results")

if (!dir.exists(results_dir)) {
  dir.create(results_dir)
}

# Read in combined scores from tp53-nf1-classifier 
classifier_score_initial <- read_tsv(file.path(results_dir, "combined_scores.tsv")) %>%
  dplyr::rename(Kids_First_Biospecimen_ID = sample_id) %>%
  left_join(clinical[,c("Kids_First_Biospecimen_ID", "RNA_library")], by = "Kids_First_Biospecimen_ID") %>%
  left_join(histology[,c("Kids_First_Biospecimen_ID", "match_id")]) %>%
  select(match_id, tp53_score, RNA_library) %>%
  unique() %>%
  # take mean score for cell lines (chordoma) with no passage info to distinguish
  group_by(match_id, RNA_library) %>%
  summarise(tp53_score = mean(tp53_score))

# print out the remaining duplicates here - these are samples with two RNA libraries sequenced and they generally have similar scores, even though run separately, so that is a nice validation. Let's take an average of these as well. This will get us 1 score per match ID.
nrow(classifier_score_initial)-length(unique(classifier_score_initial$match_id))

classifier_score_initial[duplicated(classifier_score_initial$match_id) | duplicated(classifier_score_initial$match_id, fromLast = TRUE), ]

classifier_score <- classifier_score_initial %>%
  group_by(match_id) %>%
  summarize(tp53_score = mean(tp53_score))

```

# Copy number per sample 
# overlapping functional domain
```{r CNV overlaps}
cnv_domain_overlap <- read_tsv(file.path(results_dir, "loss_overlap_domains_tp53.tsv")) %>%
  dplyr::rename(Kids_First_Biospecimen_ID=biospecimen_id) %>%
  left_join(histology[,c("Kids_First_Biospecimen_ID", "match_id")], by = "Kids_First_Biospecimen_ID") %>%
  select(match_id, copy_number, ploidy, domain) %>%
  unique()

# check duplicates - These are coming from two deceased samples which come from different brain regions we do not yet have annotated, so best to not combine.
nrow(cnv_domain_overlap)-length(unique(cnv_domain_overlap$match_id))

cnv_domain_overlap[duplicated(cnv_domain_overlap$match_id) | duplicated(cnv_domain_overlap$match_id, fromLast = TRUE), ]
```

# Structural variant overlapping 
# or within gene locus of TP53
```{r SV overlaps}
sv_overlap <- read_tsv(file.path(results_dir, "sv_overlap_tp53.tsv")) %>%
  select(-match_id) %>% 
  left_join(histology[,c("Kids_First_Biospecimen_ID", "match_id")], by = "Kids_First_Biospecimen_ID") %>%
  select(match_id, Gene_name, SV_chrom, SV_start, SV_end, SV_length, SV_type) %>%
  unique() %>%
  # collapse SVs per match id
  group_by(match_id) %>%
  summarize(SVs = str_c(paste0(SV_type, "=", SV_start, "-", SV_end), collapse = ";"),
            SV_type = str_c(unique(SV_type), collapse = ";")) %>%
  mutate(SV_counts = stringr::str_count(SVs, "="))

# check for duplicate match ids  
nrow(sv_overlap)-length(unique(sv_overlap$match_id))

sv_overlap[duplicated(sv_overlap$match_id) | duplicated(sv_overlap$match_id, fromLast = TRUE), ]
```

# Structural variant as a fusion in exon1/intron1
```{r fusions}
fusion_overlap <- read_tsv(file.path(results_dir,"fusion_bk_tp53_loss.tsv")) %>%
  left_join(histology[,c("Kids_First_Biospecimen_ID", "match_id")], by = "Kids_First_Biospecimen_ID") %>%
  select(match_id, FusionName) %>%
  group_by(match_id) %>%
  summarize(Fusions = str_c(unique(FusionName), collapse = ";")) %>%
  mutate(Fusion_counts = stringr::str_count(Fusions, "--"))

# check for duplicate match ids  
nrow(fusion_overlap)-length(unique(fusion_overlap$match_id))

fusion_overlap[duplicated(fusion_overlap$match_id) | duplicated(fusion_overlap$match_id, fromLast = TRUE), ]

```

### Check if all function sites are in hotspots

All functional sites are just 1 base so checking in `hotspot_database_2017_snv`

```{r}

functional_sites_tp53 %>%
  filter(!gsub("[A-Z|a-z]","",p53) %in% hotspot_database_2017_tp53_snv$Amino_Acid_Position )

```
2 functional sites are missing in hotspots database.

## SNV for TP53

Removing Silent or Intron classified variants to capture only putative damaging mutations

```{r}

tumor_only_snv <- data.table::fread(file.path(data_dir, "snv-mutect2-tumor-only-plus-hotspots.maf.tsv.gz"), 
                                    select = c("Chromosome",
                                              "Start_Position",
                                              "End_Position",
                                              "Strand",
                                              "Variant_Classification",
                                              "Tumor_Sample_Barcode",
                                              "Hugo_Symbol",
                                              "HGVSp_Short"),
                                   data.table = FALSE)

consensus_tp53_snv_indel <- data.table::fread(
  file.path(data_dir,"snv-consensus-plus-hotspots.maf.tsv.gz"),
                                   select = c("Chromosome",
                                              "Start_Position",
                                              "End_Position",
                                              "Strand",
                                              "Variant_Classification",
                                              "Tumor_Sample_Barcode",
                                              "Hugo_Symbol",
                                              "HGVSp_Short"),
                                   data.table = FALSE) %>%
  bind_rows(tumor_only_snv) %>% 
  dplyr::rename(Kids_First_Biospecimen_ID = Tumor_Sample_Barcode) %>%
  filter(Hugo_Symbol == "TP53") %>% 
  filter(!(Variant_Classification %in% c("Silent", "Intron",
                                         # remove other non-amino acid SNVs
                                         "3'Flank" ,"5'Flank",
                                         "3'UTR", "5'UTR"  )))
```

###  Gather annotation for SNV 

hotspots : overlaps AA position which are statistically significant SNV/Indels
activating : overlaps AA position which are found to act as gain-of-function according to literature

```{r}
# Gene location and domain overlap file 
bioMartDataPfamTp53 <- 
  readRDS(system.file("extdata", "pfamDataBioMart.RDS", package = "annoFuseData")) %>%
  dplyr::filter(hgnc_symbol=="TP53")

PF08563 <- bioMartDataPfamTp53 %>%
  filter(pfam_id == "PF08563") %>%
  select(domain_start, domain_end)

PF00870 <- bioMartDataPfamTp53 %>%
  filter(pfam_id == "PF00870") %>%
  select(domain_start, domain_end)

PF07710 <- bioMartDataPfamTp53 %>%
  filter(pfam_id == "PF07710") %>%
  select(domain_start, domain_end)

consensus_tp53_snv_indel_match <- consensus_tp53_snv_indel %>%
  left_join(histology[,c("Kids_First_Biospecimen_ID", "match_id")], by = "Kids_First_Biospecimen_ID") %>%
  select(-Kids_First_Biospecimen_ID) %>%
  mutate(
    hotspot = case_when(
      # strip REF and Variant AA to get Amino_Acid_Position in consensus maf
      (gsub("[A-Z|a-z]|[.]","",HGVSp_Short) %in% 
         # if overlaps the hotspot Amino_Acid_Position 
         hotspot_database_2017_tp53_snv$Amino_Acid_Position)  ~ 1,
      TRUE ~ 0),
    activating = case_when(
      # strip REF and Variant AA to get Amino_Acid_Position in consensus maf
      (gsub("[A-Z|a-z]|[.]","",HGVSp_Short) %in% 
         # if overlaps the activating  Amino_Acid_Position 
         c("273","248"))  ~ 1,
      TRUE ~ 0
    ),
    overlap_domain = case_when(Variant_Classification %in% c("Frame_Shift_Del", "In_Frame_Del", "Nonsense_Mutation") & PF08563$domain_start >= Start_Position & Start_Position <= PF08563$domain_end ~ 1,
                               Variant_Classification %in% c("Frame_Shift_Del", "In_Frame_Del", "Nonsense_Mutation") & PF00870$domain_start >= Start_Position & Start_Position <= PF00870$domain_end ~ 1,
                               Variant_Classification %in% c("Frame_Shift_Del", "In_Frame_Del", "Nonsense_Mutation") & PF07710$domain_start >= Start_Position & Start_Position <= PF07710$domain_end ~ 1,
                                       TRUE ~ 0)
  ) 

```

## Gather SNV,CNV, classifier and cancer_predisposition values

We will gather values per sample_id, composition, and tumor descriptor to match DNA and RNA


```{r}

tp53_alterations <- histology %>%
  select(-Kids_First_Biospecimen_ID) %>%
  unique() %>%
  # join consensus calls overlapping hotspots
  left_join(consensus_tp53_snv_indel_match) %>%
  # join filtered cnv losses
  left_join(cnv_domain_overlap) %>%
  # join SV 
  left_join(sv_overlap) %>%
  # join fusion
  left_join(fusion_overlap) %>%
  left_join(classifier_score) %>%
  # select useful columns
  dplyr::select(match_id, tp53_score, cancer_predispositions, HGVSp_Short, names(cnv_domain_overlap), names(sv_overlap), names(fusion_overlap), hotspot, activating, overlap_domain) %>%
  arrange(match_id) %>%
  unique() %>%
  replace_na(list(hotspot = 0, activating = 0, overlap_domain = 0))

```

Convert the above to wide format

```{r}

tp53_alterations_score_wide <- tp53_alterations %>%
  mutate(cn_less_ploidy = ifelse(copy_number < ploidy, "yes", "no")) %>%
  # group 
  group_by(match_id,
           cancer_predispositions,
           tp53_score,
           SV_type,
           SV_counts,
           Fusions,
           Fusion_counts,
           cn_less_ploidy) %>%
  # 
  # match_id as rows add SNV counts,CNV loss counts 
  # and hotspot/activating mutation annotation  
  summarise(
    # summarize length of SNV and CNV alterations 
   SNV_indel_counts= length(unique(HGVSp_Short[!is.na(HGVSp_Short)])), 
    CNV_loss_counts = length(unique(copy_number[!is.na(copy_number)])),
    HGVSp_Short = toString(unique(HGVSp_Short)),
    CNV_loss_evidence = toString(unique(copy_number)),
    # summarize unique hotspot values per sample_id
    hotspot = max(unique(hotspot[!is.na(hotspot) ])),
    activating = max(unique(activating[!is.na(activating)])),
    overlap_domain = max(unique(overlap_domain[!is.na(overlap_domain) ]))) 

tp53_alterations_score_wide

```

### Add annotation

As discussed above we want to annotate TP53 mutants that are putative loss-of-function OR  gain-of-function.

```{r}

tp53_alterations_score_wide <- tp53_alterations_score_wide %>%
  mutate(
    # add tp53_altered annotation column
    tp53_altered = 
      case_when(
        # when activating == 0
        activating == 0 &
          # check if mutated variant AA position overlaps hotspot database 
          ( hotspot == 1  |
              # check if sample_id has SNV+(CNV|SV) mutation
              # suggesting both alleles are mutated
              (SNV_indel_counts >= 1 & 
                 (CNV_loss_counts >=1 | SV_counts >=1) ) |
              # check if more than 1 SNV is present
              # suggesting both alleles are mutated
              SNV_indel_counts > 1 |
              # framaeshift del or nonsense SNV overlaps a functional domain
              (overlap_domain >=1) |
              # check if SNV mutant and has
              # Li-Fraumeni syndrome suggesting
              # germline TP53 mutant as cancer predisposition 
              (SNV_indel_counts >= 1 & 
                             grepl("Li-Fraumeni syndrome",cancer_predispositions)) |
              # check if CNV|SV mutant  and has
              # Li-Fraumeni syndrome suggesting
              # germline TP53 mutant as cancer predisposition 
              ((CNV_loss_counts >= 1 | SV_counts >=1 ) & 
                             grepl("Li-Fraumeni syndrome",cancer_predispositions)) |
              # check if tp53 inactivating score for RNA_Seq is greater that 0.5
              # and has Li-Fraumeni syndrome suggesting
              # germline TP53 mutant as cancer predisposition 
              (tp53_score > 0.5 & 
                             grepl("Li-Fraumeni syndrome",cancer_predispositions)) |
              # check if tp53 inactivating score for RNA_Seq is greater that 0.5
              # and has 1 or more SNV| (CNV|SV) loss in TP53
              (tp53_score > 0.5 & 
                             (SNV_indel_counts >= 1 | (CNV_loss_counts >=1 | SV_counts >=1 )))
          ) ~ "loss",
        # when activating == 1
        activating == 1 ~ "activated",
        # if no evidence supports TP53 inactivation or activation 
        TRUE ~ "Other"
      )
  )

# check for duplicate match ids  
nrow(tp53_alterations_score_wide)-length(unique(tp53_alterations_score_wide$match_id))

tp53_alterations_score_wide[duplicated(tp53_alterations_score_wide$match_id) | duplicated(tp53_alterations_score_wide$match_id, fromLast = TRUE), ]



```

## Explore distribution of tp53_altered status vs tp53 inactivation scores

```{r, out.width="50%"}
ggplot(tp53_alterations_score_wide, aes(x = factor(tp53_altered), y = tp53_score)) +
  geom_violin()+
  geom_jitter(alpha = 0.5, width = 0.2) +
  stat_compare_means() +
  theme_bw() +
  ggtitle("Distribution of scores across tp53 altered status") +
  theme(axis.text.x = element_text(angle = 60, hjust = 1))+
  xlab("tp53 altered status") 

```
TP53 mutants with gain-of-function (activating) mutants promote tumorigenesis according to literature. [Reference](https://pubmed.ncbi.nlm.nih.gov/17417627/) and [reference](https://pubmed.ncbi.nlm.nih.gov/24677579/) have similar distribution of classifier scores to that of potential loss-of-function (multi-allelic) Tp53 mutations. 

"Other" annotated samples either don't have any high-confidence loss/gain TP53 SNVs nor CNV losses OR DNA sample is not available.


Interesting 2 bs ids annotated with Li-Fraumeni synfrome pre-disposition have very low tp53 classifier scores.

```{r}

tp53_alterations_score_wide %>%
  filter(grepl("Li-Fraumeni syndrome", cancer_predispositions),
         tp53_score < 0.5) 
  
```


## Save file 

Adding biospecimen ids to `tp53_alterations_score_wide` and saving

```{r}

tp53_alterations_score_wide %>% 
  left_join(histology[,c("Kids_First_Biospecimen_ID", "match_id")], by = "match_id") %>%
  left_join(clinical[,c("Kids_First_Biospecimen_ID", "sample_id")], by = "Kids_First_Biospecimen_ID") %>%
  select("Kids_First_Biospecimen_ID", "match_id", "sample_id", names(tp53_alterations_score_wide)) %>%
  arrange(match_id) %>%
  write_tsv(file.path(results_dir,"tp53_altered_status.tsv"))

```
